PDA-49 Fixing tests
