+++ src/test/java/org/entando/plugins/pda/core/engine/FakeEngine.java
@@ -2,6 +2,7 @@ package org.entando.plugins.pda.core.engine;
 
 import lombok.Builder;
 import org.entando.plugins.pda.core.service.group.FakeGroupService;
+import org.entando.plugins.pda.core.service.process.FakeProcessFormService;
 import org.entando.plugins.pda.core.service.process.FakeProcessService;
 import org.entando.plugins.pda.core.service.task.FakeTaskCommentService;
 import org.entando.plugins.pda.core.service.task.FakeTaskDefinitionService;
@@ -15,7 +16,9 @@ public class FakeEngine extends Engine {
     @Builder
     public FakeEngine(FakeTaskService taskService, FakeTaskDefinitionService taskDefinitionService,
             FakeTaskCommentService taskCommentService, FakeProcessService processService,
+            FakeProcessFormService processFormService,
             FakeGroupService groupService) {
-        super(TYPE, taskService, taskDefinitionService, taskCommentService, processService, groupService);
+        super(TYPE, taskService, taskDefinitionService, taskCommentService, processService, processFormService,
+                groupService);
     }
 }
